requirements.txt
redis==2.10.1
docker-py==1.1.0
urllib3==1.8
msgpack-python==0.4.6
jsontree==0.4.3
